requirements.txt
boto3==1.5.14
terminaltables==3.1.0
progressbar2==3.34.3